[refactor] webui: extract SiteHeader and SiteFooter from layout
Split the layout's header and footer into dedicated components and
collapse the repeated footer-link boilerplate into a data-driven list
