+++ webui/app/src/lib/components/index.ts
@@ -1,1 +1,3 @@
 export { default as StatusMessage } from './StatusMessage.svelte';
+export { default as SiteHeader } from './SiteHeader.svelte';
+export { default as SiteFooter } from './SiteFooter.svelte';
